worker/requirements.txt
psycopg2-binary==2.9.9
httpx==0.27.2
apscheduler==3.10.4
